+++ lib/sql_commander_mysql.dart
@@ -1,5 +1,5 @@
 /// SQL Commander for MySQL library
-library sql_commander;
+library sql_commander.mysql;
 
 export 'sql_commander.dart';
 export 'src/db_mysql.dart';