+++ lib/sql_commander_postgres.dart
@@ -1,5 +1,5 @@
 /// SQL Commander for PostgreSQL library
-library sql_commander;
+library sql_commander.postgres;
 
 export 'sql_commander.dart';
 export 'src/db_postgres.dart';